Adds Spec.Automation section for the FluxHelmRelease custom resource
Helm Charts can store image information in a variety of ways:
- maps with one or more key levels
- image information provided through one or more fields

The Spec.Automation section is a map with keys being an
image (domain/image) and values indicating where, within the
Spec.Values, to find the image tag. The map (key per image)
is to be provided if automation is ever required.

The implementation will allow to easily extract image information and
makes it possible to update image tag whatever the Chart values
format is.